+++ django_mysql/models/aggregates.py
@@ -67,7 +67,8 @@ if django.VERSION < (1, 8):
 
     class SQLGroupConcat(SQLAggregate):
 
-        def __init__(self, col, distinct=False, separator=None, **extra):
+        def __init__(self, col, distinct=False, separator=None,
+                     ordering=None, **extra):
             super(SQLGroupConcat, self).__init__(col, **extra)
 
             self.distinct = distinct
@@ -75,6 +76,11 @@ if django.VERSION < (1, 8):
 
             # This can/will be improved to SetTextField or ListTextField
             self.field = CharField()
+
+            if ordering not in ('asc', 'desc', None):
+                raise ValueError(
+                    "'ordering' must be one of 'asc', 'desc', or None")
+            self.ordering = ordering
 
         sql_function = 'GROUP_CONCAT'
 
@@ -90,7 +96,10 @@ if django.VERSION < (1, 8):
             template.append("%(field)s")
 
             if self.separator is not None:
-                template.append(' SEPARATOR "{}"'.format(self.separator))
+                template.append(" SEPARATOR '{}'".format(self.separator))
+
+            if self.ordering is not None:
+                template.append(" ORDER BY %(field)s " + self.ordering.upper())
 
             template.append(")")
 
@@ -115,7 +124,7 @@ else:
         function = 'GROUP_CONCAT'
 
         def __init__(self, expression, distinct=False, separator=None,
-                     **extra):
+                     ordering=None, **extra):
 
             if 'output_field' not in extra:
                 # This can/will be improved to SetTextField or ListTextField
@@ -126,20 +135,37 @@ else:
             self.distinct = distinct
             self.separator = separator
 
-        @cached_property
-        def template(self):
-            # Constructing a template...
+            if ordering not in ('asc', 'desc', None):
+                raise ValueError(
+                    "'ordering' must be one of 'asc', 'desc', or None")
+            self.ordering = ordering
 
-            template = ["%(function)s("]
-
+        def as_sql(self, compiler, connection, function=None, template=None):
+            connection.ops.check_expression_support(self)
+            sql = ["GROUP_CONCAT("]
             if self.distinct:
-                template.append("DISTINCT ")
+                sql.append("DISTINCT ")
 
-            template.append("%(field)s")
+            expr_parts = []
+            params = []
+            for arg in self.source_expressions:
+                arg_sql, arg_params = compiler.compile(arg)
+                expr_parts.append(arg_sql)
+                params.extend(arg_params)
+            expr_sql = self.arg_joiner.join(expr_parts)
+
+            sql.append(expr_sql)
 
             if self.separator is not None:
-                template.append(' SEPARATOR "{}"'.format(self.separator))
+                sql.append(" SEPARATOR '{}'".format(self.separator))
 
-            template.append(")")
+            if self.ordering is not None:
+                sql.append(" ORDER BY ")
+                sql.append(expr_sql)
+                params.extend(params[:])
+                sql.append(" ")
+                sql.append(self.ordering.upper())
 
-            return "".join(template)
+            sql.append(")")
+
+            return "".join(sql), params